Restore fc40 specfile and add build conditions for xcp-ng
This patch restores the original fc40 specfile. It keeps the XS and
xcp-ng modifications enclosed by %bcond directives. The xcpng bcond
enables patches (fixes, CVEs) and compilation options specific to xcp-ng
build. The build_minimal bcond is used to enable/disable build of
libcurl-minimal package (disabled for an xcp-ng build).

This specfile produces the same build as the previous commit (with one
slight difference about --with-libssh2 which is now really enabled) so
its release number is not bumped.

Changelog entries will be added when the merge for the next Curl version
update will be done.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>